Added initial docker build support
Created Dockerfile that builds a Docker image with paraply2 and its
dependencies bundled. It uses the official Node docker image which does
the grunt work of copying the app inside the container.

Expose port 8888, the default paraply2 port, and the
/usr/src/app/config folder to make configuration easy.

Added .dockerignore file to not package sensitive files and files that
are not required for running.

Added config/config.coffee to .gitignore.

Todo:
Add info about Docker setup in README.md.
Add docker-compose file.