Add working docker config to project
Added working Docker config to project.

The Dockerfile declares a custom entrypoint.sh to write the Paraply2
configuration. It also resets the CMD to npm start.

Added config.tmp file that is used for app configuration when running
with docker.

The entry point file uses sed and environment variables that are set by
the docker command to replace template values with the proper values
for the config.coffee file.

The docker-compose.yml.dist shows you how to start paraply2 and
elasticsearch with docker-compose.

Updated app/db.coffee to retry the db connection as the elasticsearch
container might now be finished starting the elasticsearch server when
the paraply2 app starts.